Add get by id and post endpoints

+++ Services/EntriesService.cs
@@ -1,4 +1,5 @@
 using LxApi.Models;
+using MongoDB.Bson;
 using MongoDB.Driver;
 
 namespace LxApi.Services;
@@ -14,11 +15,20 @@ public class EntriesService : IEntriesService {
         return await _entryCollection.Find(_ => true).ToListAsync();
     }
 
+    public async Task<Entry?> GetByIdAsync(string id) {
+        if(ObjectId.TryParse(id, out _)) {
+            return await _entryCollection.Find(entry => entry.Id == id).FirstOrDefaultAsync();
+        }
+        return null;
+    }
+
     public async Task CreateAsync(Entry entry) {
         await _entryCollection.InsertOneAsync(entry);
     }
 
     public async Task DeleteAsync(string id) {
-        await _entryCollection.DeleteOneAsync(x => x.Id == id);
+        if(ObjectId.TryParse(id, out _)) {
+            await _entryCollection.DeleteOneAsync(x => x.Id == id);
+        }
     }
 }
